refactor: import cell geometry from _cellkit directly
``_cardfit`` re-exported eight of ``_cellkit``'s names behind a
``# noqa: F401`` so the card widgets could take fitting and geometry off
one import line. The convenience cost the modules their boundary: a
reader of ``entity.py`` could not tell which module owns ``cell_box``,
and ``_cardfit``'s interface advertised geometry it does not implement.

Widgets now import cell geometry from the module that owns it, and
``_cardfit``'s interface is fitting alone — it keeps only the two names
it calls itself. (``_gauge`` was named alongside it in the plan but
carries no such block; it already imported ``_cellkit`` directly.)

Pure import motion, kept to the moved names so the diff stays reviewable:
765 rendered fragments are unchanged.
