+++ custom_components/geekmagic/widgets/clock.py
@@ -14,12 +14,10 @@ from ._card import card_html, chip_html
 from ._cardfit import (
     HERO_SHARE_SOLO,
     HERO_SHARE_STACKED,
-    cell_box,
-    chip_band_px,
     fit_hero,
     hero_block,
-    label_px,
 )
+from ._cellkit import cell_box, chip_band_px, label_px
 from .base import Widget, WidgetConfig
 
 if TYPE_CHECKING:
