+++ custom_components/geekmagic/widgets/icon.py
@@ -6,7 +6,8 @@ from html import escape
 from typing import TYPE_CHECKING, ClassVar
 
 from ..htmldoc import css_rgb, mdi_span
-from ._cardfit import cell_box, fit_caption_sized, label_px
+from ._cardfit import fit_caption_sized
+from ._cellkit import cell_box, label_px
 from .base import Widget, WidgetConfig
 
 if TYPE_CHECKING:
